+++ docs/conf.py
@@ -66,6 +66,10 @@ extensions = [
     'sphinx.ext.viewcode',   # disabed, raises anexception
     'sphinx.ext.ifconfig',
     'sphinx_git',            # requires 'sphinx-git' Python package
+    # Note: sphinx_rtd_theme is not compatible with sphinxcontrib.fulltoc,
+    # but since it already provides a full TOC in the navigation pane, the
+    # sphinxcontrib.fulltoc extension is not needed.
+    'sphinx_rtd_theme',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -89,7 +93,7 @@ else:
 # General information about the project.
 project = u'zhmcclient'
 copyright = u'IBM'
-author = u'IBM Z KVM OpenStack Team'
+author = u'zhmcclient team'
 
 # The short description of the package.
 _short_description = u'Client library for IBM Z Hardware Management Console Web Services API'
@@ -222,13 +226,17 @@ viewcode_import = True
 
 # The theme to use for HTML and HTML Help pages.
 # See http://www.sphinx-doc.org/en/stable/theming.html for built-in themes.
-html_theme = "classic"
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.
 # See http://www.sphinx-doc.org/en/stable/theming.html for the options
 # available for built-in themes.
+# For options of the 'sphinx_rtd_theme', see
+# https://sphinx-rtd-theme.readthedocs.io/en/latest/configuring.html
 html_theme_options = {
+    'style_external_links': False,
+    'collapse_navigation': False,
 }
 
 # Add any paths that contain custom themes here, relative to this directory.
@@ -460,7 +468,7 @@ autodoc_mock_imports = []
 #       to datatypes of function parameters can be controlled.
 #
 intersphinx_mapping = {
-  'py': ('https://docs.python.org/2/', None), # agnostic to Python version
+  'py': ('https://docs.python.org/3/', None), # agnostic to Python version
   'py2': ('https://docs.python.org/2', None), # specific to Python 2
   'py3': ('https://docs.python.org/3', None), # specific to Python 3
 }